+++ Quantom/Quantom/Raytracer.cuh
@@ -10,8 +10,8 @@
 #include "Simulation.cuh"
 
 
-const float FOCAL_LEN_RATIO = 1;
-const float FOCAL_Y_OFFSET = -1;
+const double FOCAL_LEN_RATIO = 1;
+const double FOCAL_Y_OFFSET = -1;
 const int RAYS_PER_DIM = 1000;	// Can't launch kernels if above 1024
 const int NUM_RAYS = RAYS_PER_DIM * RAYS_PER_DIM;
 const int THREADS_PER_BLOCK = 1024;
@@ -24,10 +24,10 @@ public:
 	Ray(Float3 unit_vector, Float3 origin);
 	__device__ void reset();
 
-	__device__ bool hitsParticle(Float3* particle_center, float particle_radius);
+	__device__ bool hitsParticle(Float3* particle_center, double particle_radius);
 	__device__ void searchCompound(CompoundState* compoundstate, Box* box, int i);	// I is for bugfinding.
 	__device__ bool searchSolvent(Float3* pos, Box* box, int solvent_index);
-	__device__ float distToPoint(Float3 point);
+	__device__ double distToPoint(Float3 point);
 	
 
 	Float3 origin;
@@ -36,13 +36,13 @@ public:
 
 
 	// Changing with each render step
-	float closest_collision = 0;
+	double closest_collision = 0;
 	int atom_type = -1;
 	bool log_particle = false;
 
 
 private:
-	__device__ float distToSphereIntersect(Float3* particle_center, float particle_radius);
+	__device__ double distToSphereIntersect(Float3* particle_center, double particle_radius);
 
 	
 };
